Fix deletion of dead Session instances
The Java object holds a reference to the websocket::Session pointer in a long value.
There is a race that occurs between Java calling the Session::close() method and the
libwebsocket that is calling Session::OnDestroy() immediately after Session::OnConnectError().
It is possible that the websocket::Session pointer is released while the Java object
still has a reference to it.  One solution could be to make the call to Session::close()
in Java while having the connection lock (this would serialize the Session::close() and
deletion of websocket::Session).  The solution choosen is to delete the previous Session
instance only when we create a new websocket::Session because at this time, the Java side
has cleared every destroyed instance.