Fix providing the stats array to OnConnect and OnConnectError
Sometimes, the getStats() method could return an null stats array because the TLWebSocket
had it session instance cleared by another thread while we execute OnConnect() or OnConnectError().
Since we know the current websocket::Session, moved the getStats() implementation as a static
method and give it the websocket::Session that we have (from OnConnect() and OnConnectError()
the pointer is still valid because we are called from the service thread and it is not yet released).