Fix double delete of websocket::Session
When lws_client_connect_via_info() reaches some specific errors, it
can call immediately the callback and we execute either OnConnectError()
or OnDestroy().  Because OnConnectError() can in turn start new connections
(through the proxies), this can trigger again the OnConnectError() and
OnDestroy() callbacks but we are still creating the websocket instance.
The caller (Java, Objective-C) is not yet aware of the websocket::Session
instance.  These cases are triggered by:

- a NXDOMAIN error when resolving the hostname to an IP address.
  On Android and iOS, such error occurs when the network is being
  setting up.  When this error occurs, the OnConnectError() is called
  (we then start the connection through proxies).
- when a server IP address cannot be used (ie, no local IPv4 or
  no local IPv6), then lws_client_connect_via_info() fails and
  it seems that the OnConnectError() is not called, then lws_client_connect_via_info()
  will return null in that case.